chore: change engine to Node.js v14
The purpose of this change is to squelch a "bad engine" warning when installing under Node.js v14.

This package currently supports Node.js v14, but:

1. the `engines` field says otherwise, and
2. TS is configured to target Node.js v16

The resulting build artifact (`dist/index.js`) is _identical_ to the previous changeset.

To that end, I've changed the configuration and dependency from `@tsconfig/node16` to `@tsconfig/node14`.  I have also enabled the build to run across multiple Node.js versions and upgraded a couple actions while I was there.

It also looked like there was a problem [on npm](https://npm.im) referencing the repo, so I fiddled with the `repository` field.  Could just be old, though.